try to sync attack & hit animations

+++ Core/Sprite/MotionRequest.cs
@@ -2,6 +2,6 @@
 
 public struct MotionRequest {
 	public SpriteMotion Motion;
-	public double delay;
+	public float delay;
 	public bool forced;
 }
